microservices/shared/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
motor==3.3.1
pymongo==4.5.0
redis==7.1.1
pydantic==2.12.5
python-jose==3.5.0
passlib==1.7.4
bcrypt==4.1.3
python-dotenv==1.2.1
python-multipart==0.0.22
razorpay==2.0.0
httpx==0.28.1
